feature product

+++ src/main/java/com/onlineshop/controller/ProductController.java
@@ -25,7 +25,7 @@ public class ProductController {
             @RequestParam(required = false) String brand,
             @RequestParam(required = false) BigDecimal minPrice,
             @RequestParam(required = false) BigDecimal maxPrice,
-            @RequestParam(required = false) Integer minRating,
+            @RequestParam(required = false) Float minRating,
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "12") int size
     ) {
